+++ test/Fixture/AttributeWithParameters.php
@@ -1,0 +1,34 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Copyright (c) 2025 Andreas Möller
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE.md file that was distributed with this source code.
+ *
+ * @see https://github.com/ergebnis/attribute-collector
+ */
+
+namespace Ergebnis\AttributeCollector\Test\Fixture;
+
+#[\Attribute(\Attribute::TARGET_ALL | \Attribute::IS_REPEATABLE)]
+final class AttributeWithParameters
+{
+    public function __construct(
+        private string $foo,
+        private int $bar,
+    ) {
+    }
+
+    public function foo(): string
+    {
+        return $this->foo;
+    }
+
+    public function bar(): int
+    {
+        return $this->bar;
+    }
+}