+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
-For a full diff see [`1902cc2...main`][1902cc2...main].
+For a full diff see [`e66cfff...main`][e66cfff...main].
 
-[1902cc2...main]: https://github.com/ergebnis/attribute-collector/compare/1902cc2...main
+### Added
+
+- Added `Attribute` ([#1]), by [@localheinz]
+
+[e66cfff...main]: https://github.com/ergebnis/attribute-collector/compare/e66cfff...main
+
+[#1]: https://github.com/ergebnis/attribute-collector/pull/1
+
+[@localheinz]: https://github.com/localheinz